pyproject.toml
[project]
name = "bazarify-ansible"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "rich>=14.3.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
